separate type for JSON Serialize
